Joining the table and creating table for analysis
We're joining the table on category_id and id of the tables. This acts like a analytics layer. This code was generated using the visual glue tool. Two data catalog source from AWS athena and then in action we choose join transformation. After which I loaded the data in s3 and AWS Athena